+ refactor app.py, Notifications Feed page
